lir: split the short-circuit and match lowerings out of control.rs
control.rs held five subjects across 640 lines, past the reading budget
DOCUMENTATION.md sets. `and`/`or` and `match` are each self-contained and each
has a document of its own to cite, so they move to `control/shortcircuit.rs`
and `control/matcharms.rs` unchanged.

What is left in control.rs is the tail-argument ownership predicates, the
intrinsic specialization, `eval` and `emit`, at 354 lines.
